chore: Split utility functions by domain
